SOFT: Extract AES functions to mech_openssl.c
Extract the OpenSSL based AES functions into file mech_openssl.c,
so that they can be used by other tokens as well.
